mv dedicated ts methods from mutable state impl to time skipping path
